User weight refactoring
Refactoring user weigth from int to double

+++ app/src/main/java/com/example/fithub/AddExerciseDialogFragment.kt
@@ -118,7 +118,7 @@ class AddExerciseDialogFragment : SearchDialogFragment<ExerciseDto>() {
             try {
 
                 val user = NetworkModule.api.getUserById(userId)
-                val userWeight = user.profile.weightKg.toDouble()
+                val userWeight = user.profile.weightKg
                 val mets = exercise.mets ?: run{
                     Log.d("brak mets", "Brak danych mets w bazie")
                     return@launch
