+++ app/src/main/java/com/example/fithub/data/UpdateUserGoalDto.kt
@@ -2,8 +2,8 @@ package com.example.fithub.data
 
 data class UpdateUserGoalDto(
     val type: String? = null,
-    val firstWeightKg: Int? = null,
-    val targetWeightKg: Int? = null,
+    val firstWeightKg: Double? = null,
+    val targetWeightKg: Double? = null,
     val plan: GoalPlanData? = null,
     val status: String? = null,
     val completedAt: String? = null,
